split `auth.js` for `/pages` and `/app` dirs example
