+++ apps/sample-next-auth-app/app/in/page.tsx
@@ -1,7 +1,7 @@
 import { getServerSession } from 'next-auth'
 import { redirect } from 'next/navigation'
 import { authConfig } from '~/pages/api/auth/[...nextauth]'
-import { Profile } from '~/shared/components/profile'
+import { Profile } from '~/components/profile'
 
 export default async function In() {
   const session = await getServerSession(authConfig)
