+++ apps/sample-next-auth-app/app/page.tsx
@@ -1,6 +1,6 @@
 import { getServerSession } from 'next-auth'
 import { authConfig } from '~/pages/api/auth/[...nextauth]'
-import { Home } from '~/shared/components/home'
+import { Home } from '~/components/home'
 
 export default async function Index() {
   const session = await getServerSession(authConfig)
