+++ apps/sample-next-auth-app/app/template.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from 'react'
-import styles from '~/shared/template.module.css'
+import styles from './template.module.css'
 
 export default function RootTemplate({ children }: { children: ReactNode }) {
   return <div className={styles.wrapper}>{children}</div>
