+++ apps/sample-next-auth-app/components/footer/index.tsx
