+++ apps/sample-next-auth-pages/components/footer/index.tsx
