+++ apps/sample-next-auth-pages/pages/_app.tsx
@@ -1,14 +1,13 @@
 import type { AppProps } from 'next/app'
 import { SessionProvider } from 'next-auth/react'
-import templateStyles from '~/shared/template.module.css'
-import '~/shared/globals.css'
+import '~/lib/globals.css'
 import { Inter } from '@next/font/google'
 
 const inter = Inter({ subsets: ['latin'], display: 'swap' })
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <SessionProvider session={pageProps.session}>
-      <div className={`${templateStyles.wrapper} ${inter.className}`}>
+      <div className={`wrapper  ${inter.className}`}>
         <Component {...pageProps} />
       </div>
     </SessionProvider>
