--- apps/sample-next-auth-pages/shared/template.module.css
@@ -1,15 +1,0 @@
-@keyframes load {
-  0% {
-    transform: translateY(-2rem);
-    opacity: 0;
-  }
-
-  100% {
-    transform: translateY(0);
-    opacity: 1;
-  }
-}
-
-.wrapper {
-  animation: load 500ms ease-in;
-}
